docker-rabbitmq-gateway/requirements.txt
pika
python-dotenv==1.1.0